+++ example_using_go_router/lib/pages/page_home.dart
@@ -42,8 +42,12 @@ class _Page_HomeState extends State<Page_Home> {
       );
     });
 
+    fsdui.namedWidgets['some-named-widget'] = (context)=>_someWidget(context);
+
     super.initState();
   }
+
+  Widget _someWidget(BuildContext context) => Center(child: Text("MONKEY"));
 
   @override
   Widget build(BuildContext context) {
@@ -65,45 +69,43 @@ class _Page_HomeState extends State<Page_Home> {
     //   // ),
     // );
 
-    final uniqueTabBarName = DateTime.now().millisecondsSinceEpoch.toString();
+    // final uniqueTabBarName = DateTime.now().millisecondsSinceEpoch.toString();
     SnippetBuilder sp = SnippetBuilder(
       initialValue: ScaffoldNode(
-        name: 'home-scaffold-with-tabs',
-        // appBar: NamedPS(
-        //   propertyName: 'appBar',
-        //   child: AppBarNode(
-        //     // tabBarName: uniqueTabBarName,
-        //     bgColor: Colors.grey,
-        //     title: NamedSC(
-        //       propertyName: 'title',
-        //       child: TextNode(
-        //         text: 'my title',
-        //         tsPropGroup: TextStyleProperties(),
-        //       ),
-        //     ),
-        //     titleTextStyle: TextStyleProperties(),
-        //     actions: NamedMC(propertyName: 'actions', children: []),
-        //     leading: NamedSC(propertyName: 'leading'),
-        //     bottom: NamedPS(
-        //       propertyName: 'bottom',
-        //       // child: TabBarNode(
-        //       //   tabBarName: uniqueTabBarName,
-        //       //   labelTSPropGroup: TextStyleProperties(),
-        //       //   children: [
-        //       //     TextNode(text: 'tab 1', tsPropGroup: TextStyleProperties()),
-        //       //     TextNode(text: 'Tab 2', tsPropGroup: TextStyleProperties()),
-        //       //   ],
-        //       // ),
-        //     ),
-        //   ),
-        // ),
-        // body: NamedSC(
-        //   propertyName: 'body',
-        //   // child: TabBarViewNode(
-        //   //   tabBarName: uniqueTabBarName,
-        //   //   children: [PlaceholderNode(), PlaceholderNode()],
-        //   // ),
-        // ),
+        name: 'home-scaffold-with-tabs-99',
+        appBar: NamedPS(
+          propertyName: 'appBar',
+          child: AppBarNode(
+            // tabBarName: uniqueTabBarName,
+            bgColor: Colors.grey,
+            title: NamedSC(
+              propertyName: 'title',
+              child: TextNode(
+                text: 'My Scaffold with a TabBar',
+                tsPropGroup: TextStyleProperties(),
+              ),
+            ),
+            titleTextStyle: TextStyleProperties(),
+            actions: NamedMC(propertyName: 'actions', children: []),
+            leading: NamedSC(propertyName: 'leading'),
+            bottom: NamedPS(
+              propertyName: 'bottom',
+              child: TabBarNode(
+                labelTSPropGroup: TextStyleProperties(),
+                children: [
+                  TextNode(text: 'tab 1', tsPropGroup: TextStyleProperties()),
+                  TextNode(text: 'Tab 2', tsPropGroup: TextStyleProperties()),
+                ],
+              ),
+            ),
+          ),
+        ),
+        body: NamedSC(
+          propertyName: 'body',
+          child: TabBarViewNode(
+            children: [PlaceholderNode(), PlaceholderNode()],
+          ),
+        ),
       ),
 
       // snippetRootNode: SnippetRootNode(