+++ lib/src/snippet/snodes/tabbarview_node.mapper.dart
@@ -25,6 +25,9 @@ class TabBarViewNodeMapper extends SubClassMapperBase<TabBarViewNode> {
   static String? _$name(TabBarViewNode v) => v.name;
   static const Field<TabBarViewNode, String> _f$name =
       Field('name', _$name, opt: true);
+  static String? _$tabBarName(TabBarViewNode v) => v.tabBarName;
+  static const Field<TabBarViewNode, String> _f$tabBarName =
+      Field('tabBarName', _$tabBarName, opt: true);
   static List<SNode> _$children(TabBarViewNode v) => v.children;
   static const Field<TabBarViewNode, List<SNode>> _f$children =
       Field('children', _$children);
@@ -54,6 +57,7 @@ class TabBarViewNodeMapper extends SubClassMapperBase<TabBarViewNode> {
   @override
   final MappableFields<TabBarViewNode> fields = const {
     #name: _f$name,
+    #tabBarName: _f$tabBarName,
     #children: _f$children,
     #uid: _f$uid,
     #tags: _f$tags,
@@ -75,7 +79,9 @@ class TabBarViewNodeMapper extends SubClassMapperBase<TabBarViewNode> {
 
   static TabBarViewNode _instantiate(DecodingData data) {
     return TabBarViewNode(
-        name: data.dec(_f$name), children: data.dec(_f$children));
+        name: data.dec(_f$name),
+        tabBarName: data.dec(_f$tabBarName),
+        children: data.dec(_f$children));
   }
 
   @override
@@ -134,7 +140,7 @@ abstract class TabBarViewNodeCopyWith<$R, $In extends TabBarViewNode, $Out>
     implements SNodeCopyWith<$R, $In, $Out> {
   ListCopyWith<$R, SNode, SNodeCopyWith<$R, SNode, SNode>> get children;
   @override
-  $R call({String? name, List<SNode>? children});
+  $R call({String? name, String? tabBarName, List<SNode>? children});
   TabBarViewNodeCopyWith<$R2, $In, $Out2> $chain<$R2, $Out2>(
       Then<$Out2, $R2> t);
 }
@@ -152,14 +158,19 @@ class _TabBarViewNodeCopyWithImpl<$R, $Out>
       ListCopyWith($value.children, (v, t) => v.copyWith.$chain(t),
           (v) => call(children: v));
   @override
-  $R call({Object? name = $none, List<SNode>? children}) =>
+  $R call(
+          {Object? name = $none,
+          Object? tabBarName = $none,
+          List<SNode>? children}) =>
       $apply(FieldCopyWithData({
         if (name != $none) #name: name,
+        if (tabBarName != $none) #tabBarName: tabBarName,
         if (children != null) #children: children
       }));
   @override
   TabBarViewNode $make(CopyWithData data) => TabBarViewNode(
       name: data.get(#name, or: $value.name),
+      tabBarName: data.get(#tabBarName, or: $value.tabBarName),
       children: data.get(#children, or: $value.children));
 
   @override